FacenetTuning/facenet/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
cycler==0.11.0
cycler==0.11.0
flatbuffers==1.12.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
idna==3.3
imageio==2.14.0
joblib==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.19.2
oauthlib==3.2.0
opencv-python==3.4.17.63
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytest==7.1.1
psutil==5.9.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.15.0
tensorboard==2.5.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-gpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf-slim==1.1.0
threadpoolctl==3.1.0
typing-extensions==3.7.4
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.12.1